Faster psim / cleanup (#54)
* cleanup bash script and multi-process psim

* catch bad variables

* trying to un-hardcode the loop

* bugtesting running the loop on my user

* debugging the loop

* catch crossover odd

* missing f in fstring

* add a switch for job plotting

* bugfix plots

* symmetry fix and cleaning up print statements

* remove duplication, 8 to 9 skiprows

* add PATH to gens data

* lowercase args

* Update PueoCall_Array.sh

* change 40 to threads

I will be addressing the formatting in gensData and polar_plotter in a separate branch

+++ BiconeEvolution/current_antenna_evo_build/XF_Loop/Evolutionary_Loop/Antenna_Performance_Metric/VariablePlots.py
@@ -11,7 +11,7 @@ parser.add_argument("source", help="Name of source folder from home directory", 
 parser.add_argument("destination", help="Name of destination folder from home directory", type=str)
 parser.add_argument("numGens", help="Number of generations the code is running for", type=int)
 parser.add_argument("NPOP", help="Number of individuals in a population", type=int)
-parser.add_argument("GeoScalingFactor", help="The number by which we are scaling the antenna dimensions", type=int)
+parser.add_argument("GeoScalingFactor", help="The number by which we are scaling the antenna dimensions", type=float)
 g = parser.parse_args()
 
 # The name of the plot that will be put into the destination folder, g.destination
